Migrate invite_4_claimer tests
