+++ libparsec/crates/protocol/tests/invite.rs
@@ -472,83 +472,3 @@ fn serde_invite_3b_greeter_signify_trust_rep(
 
     assert_eq!(data2, expected);
 }
-
-#[parsec_test]
-fn serde_invite_4_claimer_communicate_req() {
-    // Generated from Python implementation (Parsec v2.6.0+dev)
-    // Content:
-    //   cmd: "invite_4_claimer_communicate"
-    //   payload: hex!("666f6f626172")
-    let raw = hex!(
-        "82a3636d64bc696e766974655f345f636c61696d65725f636f6d6d756e6963617465a77061"
-        "796c6f6164c406666f6f626172"
-    );
-
-    let req = invited_cmds::invite_4_claimer_communicate::Req {
-        payload: b"foobar".as_ref().into(),
-    };
-
-    let expected = invited_cmds::AnyCmdReq::Invite4ClaimerCommunicate(req);
-
-    let data = invited_cmds::AnyCmdReq::load(&raw).unwrap();
-
-    assert_eq!(data, expected);
-
-    // Also test serialization round trip
-    let raw2 = if let invited_cmds::AnyCmdReq::Invite4ClaimerCommunicate(data) = data {
-        data.dump().unwrap()
-    } else {
-        unreachable!()
-    };
-
-    let data2 = invited_cmds::AnyCmdReq::load(&raw2).unwrap();
-
-    assert_eq!(data2, expected);
-}
-
-#[parsec_test]
-#[case::ok(
-    // Generated from Python implementation (Parsec v2.6.0+dev)
-    // Content:
-    //   payload: hex!("666f6f626172")
-    //   status: "ok"
-    &hex!(
-        "82a77061796c6f6164c406666f6f626172a6737461747573a26f6b"
-    )[..],
-    invited_cmds::invite_4_claimer_communicate::Rep::Ok {
-        payload: b"foobar".as_ref().into(),
-    }
-)]
-#[case::not_found(
-    // Generated from Python implementation (Parsec v2.6.0+dev)
-    // Content:
-    //   status: "not_found"
-    &hex!(
-        "81a6737461747573a96e6f745f666f756e64"
-    )[..],
-    invited_cmds::invite_4_claimer_communicate::Rep::NotFound
-)]
-#[case::invalid_state(
-    // Generated from Python implementation (Parsec v2.6.0+dev)
-    // Content:
-    //   status: "invalid_state"
-    &hex!(
-        "81a6737461747573ad696e76616c69645f7374617465"
-    )[..],
-    invited_cmds::invite_4_claimer_communicate::Rep::InvalidState
-)]
-fn serde_invite_4_claimer_communicate_rep(
-    #[case] raw: &[u8],
-    #[case] expected: invited_cmds::invite_4_claimer_communicate::Rep,
-) {
-    let data = invited_cmds::invite_4_claimer_communicate::Rep::load(raw).unwrap();
-
-    assert_eq!(data, expected);
-
-    // Also test serialization round trip
-    let raw2 = data.dump().unwrap();
-
-    let data2 = invited_cmds::invite_4_claimer_communicate::Rep::load(&raw2).unwrap();
-
-    assert_eq!(data2, expected);
-}
