+++ libparsec/crates/protocol/tests/invited_cmds/v3/invite_4_claimer_communicate.rs
@@ -1,27 +1,129 @@
 // Parsec Cloud (https://parsec.cloud) Copyright (c) BUSL-1.1 2016-present Scille SAS
 
+// `allow-unwrap-in-test` don't behave as expected, see:
+// https://github.com/rust-lang/rust-clippy/issues/11119
+#![allow(clippy::unwrap_used)]
+
+use libparsec_tests_lite::prelude::*;
+use libparsec_types::prelude::*;
+
 use super::invited_cmds;
 
 // Request
 
 pub fn req() {
-    // TODO #4545: Implement test
+    // Generated from Python implementation (Parsec v2.6.0+dev)
+    // Content:
+    //   cmd: "invite_4_claimer_communicate"
+    //   payload: hex!("666f6f626172")
+    let raw = hex!(
+        "82a3636d64bc696e766974655f345f636c61696d65725f636f6d6d756e6963617465a77061"
+        "796c6f6164c406666f6f626172"
+    );
+
+    let req = invited_cmds::invite_4_claimer_communicate::Req {
+        payload: b"foobar".as_ref().into(),
+    };
+
+    let expected = invited_cmds::AnyCmdReq::Invite4ClaimerCommunicate(req);
+
+    let data = invited_cmds::AnyCmdReq::load(&raw).unwrap();
+
+    p_assert_eq!(data, expected);
+
+    // Also test serialization round trip
+    let invited_cmds::AnyCmdReq::Invite4ClaimerCommunicate(req2) = data else {
+        unreachable!()
+    };
+
+    let raw2 = req2.dump().unwrap();
+
+    let data2 = invited_cmds::AnyCmdReq::load(&raw2).unwrap();
+
+    p_assert_eq!(data2, expected);
 }
 
 // Responses
 
 pub fn rep_ok() {
-    // TODO #4545: Implement test
+    // Generated from Python implementation (Parsec v2.6.0+dev)
+    // Content:
+    //   payload: hex!("666f6f626172")
+    //   status: "ok"
+    let raw = hex!("82a77061796c6f6164c406666f6f626172a6737461747573a26f6b");
+
+    let expected = invited_cmds::invite_4_claimer_communicate::Rep::Ok {
+        payload: b"foobar".as_ref().into(),
+    };
+
+    let data = invited_cmds::invite_4_claimer_communicate::Rep::load(&raw).unwrap();
+
+    p_assert_eq!(data, expected);
+
+    // Also test serialization round trip
+    let raw2 = data.dump().unwrap();
+
+    let data2 = invited_cmds::invite_4_claimer_communicate::Rep::load(&raw2).unwrap();
+
+    p_assert_eq!(data2, expected);
 }
 
 pub fn rep_already_deleted() {
-    // TODO #4545: Implement test
+    // Generated from Rust implementation (Parsec v2.15.0+dev)
+    // Content:
+    //   status: "already_deleted"
+    let raw = hex!("81a6737461747573af616c72656164795f64656c65746564");
+
+    let expected = invited_cmds::invite_1_claimer_wait_peer::Rep::AlreadyDeleted;
+
+    let data = invited_cmds::invite_1_claimer_wait_peer::Rep::load(&raw).unwrap();
+
+    p_assert_eq!(data, expected);
+
+    // Also test serialization round trip
+    let raw2 = data.dump().unwrap();
+
+    let data2 = invited_cmds::invite_1_claimer_wait_peer::Rep::load(&raw2).unwrap();
+
+    p_assert_eq!(data2, expected);
 }
 
 pub fn rep_not_found() {
-    // TODO #4545: Implement test
+    // Generated from Python implementation (Parsec v2.6.0+dev)
+    // Content:
+    //   status: "not_found"
+    let raw = hex!("81a6737461747573a96e6f745f666f756e64");
+
+    let expected = invited_cmds::invite_4_claimer_communicate::Rep::NotFound;
+
+    let data = invited_cmds::invite_4_claimer_communicate::Rep::load(&raw).unwrap();
+
+    p_assert_eq!(data, expected);
+
+    // Also test serialization round trip
+    let raw2 = data.dump().unwrap();
+
+    let data2 = invited_cmds::invite_4_claimer_communicate::Rep::load(&raw2).unwrap();
+
+    p_assert_eq!(data2, expected);
 }
 
 pub fn rep_invalid_state() {
-    // TODO #4545: Implement test
+    // Generated from Python implementation (Parsec v2.6.0+dev)
+    // Content:
+    //   status: "invalid_state"
+    let raw = hex!("81a6737461747573ad696e76616c69645f7374617465");
+
+    let expected = invited_cmds::invite_4_claimer_communicate::Rep::InvalidState;
+
+    let data = invited_cmds::invite_4_claimer_communicate::Rep::load(&raw).unwrap();
+
+    p_assert_eq!(data, expected);
+
+    // Also test serialization round trip
+    let raw2 = data.dump().unwrap();
+
+    let data2 = invited_cmds::invite_4_claimer_communicate::Rep::load(&raw2).unwrap();
+
+    p_assert_eq!(data2, expected);
 }
